+++ lib/remote_control_webrtc.dart
@@ -1,0 +1,8 @@
+
+import 'remote_control_webrtc_platform_interface.dart';
+
+class RemoteControlWebrtc {
+  Future<String?> getPlatformVersion() {
+    return RemoteControlWebrtcPlatform.instance.getPlatformVersion();
+  }
+}