+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2026-09-01
+
+- Upgrade Flutter plugin project structure with MethodChannel and Android support
+- Upgrade Flutter and Android dependencies
+
 ## [1.0.3] - 2026-07-06
 
 - modify dependencies, and add flags for Kotlin and new DSL